+++ SMASHING/Global/Components/BlueCTAButton.swift
@@ -1,5 +1,5 @@
 //
-//  CTAButton.swift
+//  BlueCTAButton.swift
 //  SMASHING
 //
 //  Created by 이승준 on 1/7/26.
