+++ SMASHING/Presentation/Review/TierExplanationView.swift
@@ -12,11 +12,11 @@ import Then
 
 final class TierExplanationView: BaseUIView {
     
-    //MARK: - Properties
+    // MARK: - Properties
     
     var dismissAction: (() -> Void)?
     
-    //MARK: - UI Components
+    // MARK: - UI Components
     
     private lazy var navigationBar = CustomNavigationBar(title: "티어 설명").then {
         $0.setRightButton(image: UIImage.icCloseLg, action: dismissAction ?? {})
@@ -99,7 +99,7 @@ final class TierExplanationView: BaseUIView {
         return collection
     }()
     
-    //MARK: - Lifecycle
+    // MARK: - Lifecycle
     
     override func setUI() {
         addSubviews(navigationBar, imageView, tierLabel, detailStack, tierCollectionView,
@@ -264,7 +264,7 @@ extension TierExplanationViewController: UICollectionViewDelegate, UICollectionV
         if collectionView == mainView.tierCollectionView {
             return OreTier.allCases.count
         } else {
-            return OreTier.diamond.skills(sports: .tableTennis).count
+            return oreTier.skills(sports: sports).count
         }
     }
 
@@ -298,55 +298,3 @@ extension TierExplanationViewController: UICollectionViewDelegate, UICollectionV
     }
 }
 
-final class SkillExplanationCell: BaseUICollectionViewCell, ReuseIdentifiable {
-    
-    // MARK: - UI Components
-    
-    private let containerView = UIView().then {
-        $0.backgroundColor = .Background.surface
-        $0.layer.cornerRadius = 12
-        $0.clipsToBounds = true
-    }
-    
-    private let titleLabel = UILabel().then {
-        $0.textColor = .Text.primary
-        $0.font = .pretendard(.textMdSb)
-        $0.numberOfLines = 1
-    }
-    
-    private let subtitleLabel = UILabel().then {
-        $0.textColor = .Text.secondary
-        $0.font = .pretendard(.textSmR)
-        $0.numberOfLines = 0
-        $0.lineBreakMode = .byWordWrapping
-    }
-    
-    // MARK: - Lifecycle
-    
-    override func setUI() {
-        contentView.addSubview(containerView)
-        containerView.addSubviews(titleLabel, subtitleLabel)
-    }
-    
-    override func setLayout() {
-        containerView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-        
-        titleLabel.snp.makeConstraints {
-            $0.top.equalToSuperview().inset(12)
-            $0.leading.trailing.equalToSuperview().inset(16)
-        }
-        
-        subtitleLabel.snp.makeConstraints {
-            $0.top.equalTo(titleLabel.snp.bottom).offset(8)
-            $0.leading.trailing.equalToSuperview().inset(16)
-            $0.bottom.equalToSuperview().inset(12)
-        }
-    }
-    
-    func configure(with skill: SkillExplanation) {
-        titleLabel.text = skill.name
-        subtitleLabel.text = skill.explanation
-    }
-}
